renamed a service folder to services
