+++ package.json
@@ -1,5 +1,4 @@
 {
-
   "name": "notesharing",
   "version": "0.0.0",
   "scripts": {
@@ -17,12 +16,14 @@
     "@angular/common": "~10.1.6",
     "@angular/compiler": "~10.1.6",
     "@angular/core": "~10.1.6",
+    "@angular/fire": "^6.1.4",
     "@angular/forms": "~10.1.6",
     "@angular/material": "^10.2.7",
     "@angular/platform-browser": "~10.1.6",
     "@angular/platform-browser-dynamic": "~10.1.6",
     "@angular/router": "~10.1.6",
     "angular-bootstrap-md": "^10.1.1",
+    "firebase": "^8.1.2",
     "modules": "^0.4.0",
     "ngx-mat-select-search": "^3.1.4",
     "node": "^15.0.1",
@@ -50,5 +51,4 @@
     "tslint": "~6.1.0",
     "typescript": "~4.0.2"
   }
-
 }
