+++ src/app/components/signup/signup.component.ts
@@ -1,4 +1,4 @@
-import { AuthenticateService } from './../../service/authenticateservice.service';
+import { AuthenticateService } from './../../services/authenticateservice.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators, FormControl } from '@angular/forms';
 import { Router } from '@angular/router';
